+++ x/tokenconverter/module/depinject.go
@@ -27,10 +27,11 @@ func init() {
 type ModuleInputs struct {
 	depinject.In
 
-	Config       *types.Module
-	Environment  appmodule.Environment
-	Cdc          codec.Codec
-	AddressCodec address.Codec
+	Config                *types.Module
+	Environment           appmodule.Environment
+	Cdc                   codec.Codec
+	AddressCodec          address.Codec
+	ValidatorAddressCodec address.ValidatorAddressCodec
 
 	AccountKeeper      types.AccountKeeper
 	BankKeeper         types.BankKeeper
@@ -55,6 +56,7 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 		in.Environment,
 		in.Cdc,
 		in.AddressCodec,
+		in.ValidatorAddressCodec,
 		authority,
 		in.AccountKeeper,
 		in.BankKeeper,
